Remove `ConsoleFragmentLogger` and add `ConsoleLogger` package trait
