+++ Package.swift
@@ -12,6 +12,10 @@ let package = Package(
     products: [
         .library(name: "ConsoleKit", targets: ["ConsoleKit"])
     ],
+    traits: [
+        "ConsoleLogger",
+        .default(enabledTraits: ["ConsoleLogger"]),
+    ],
     dependencies: [
         .package(url: "https://github.com/apple/swift-log.git", from: "1.6.3"),
         .package(url: "https://github.com/apple/swift-async-algorithms.git", from: "1.0.4"),
@@ -20,7 +24,7 @@ let package = Package(
         .target(
             name: "ConsoleKit",
             dependencies: [
-                .product(name: "Logging", package: "swift-log"),
+                .product(name: "Logging", package: "swift-log", condition: .when(traits: ["ConsoleLogger"])),
                 .product(name: "AsyncAlgorithms", package: "swift-async-algorithms"),
             ],
             swiftSettings: swiftSettings
