+++ Sources/ConsoleKit/Docs.docc/index.md
@@ -12,7 +12,7 @@
 
 * Utilities for sending text (including styles and colors, when supported) to a terminal.
 * Utilities for reading input from a terminal.
-* ``ConsoleLogger`` and ``ConsoleFragmentLogger``, [SwiftLog](https://github.com/apple/swift-log) `LogHandler` implementations for customizable logging to a console.
+* ``ConsoleLogger``, a [SwiftLog](https://github.com/apple/swift-log) `LogHandler` implementation for customizable logging to a console.
 
 ## Topics
 
@@ -42,7 +42,6 @@
 ### Logging
 
 - ``ConsoleLogger``
-- ``ConsoleFragmentLogger``
 - ``LoggerFragment``
 - ``LogRecord``
 - ``FragmentOutput``
